Refacto script to generate the changelog
